+++ src/rest/index.ts
@@ -1,2 +1,3 @@
 export { RestApplication } from './rest.application.js';
+export { Route } from './rest.const.js';
 export { createRestApplicationContainer } from './rest.container.js';
