+++ src/rest/rest.const.ts
@@ -1,0 +1,7 @@
+export enum Route {
+  Users = '/users',
+  Offers = '/offers',
+  Reviews = '/reviews',
+  Upload = '/upload',
+  Static = '/static'
+}
